RG&MM - updated more xlsx to document
