+++ ldk/javascript/examples/self-test-loop/src/config/document-group.ts
@@ -1,13 +1,13 @@
 import TestGroup from '../testingFixtures/testGroup';
 import { LoopTest } from '../testingFixtures/loopTest';
 
-import * as xlsxTests from '../tests/xlsx';
+import * as documentTests from '../tests/document';
 
-export const xlsxTestGroup = (): TestGroup =>
-  new TestGroup('XLSX Aptitude', [
+export const documentTestGroup = (): TestGroup =>
+  new TestGroup('Document Aptitude', [
     new LoopTest(
       'Excel Aptitude - Encode test',
-      xlsxTests.testXLSXEncodeAndDecode,
+      documentTests.testDocumentEncodeAndDecode,
       5000,
       'Encodes a workbook object into Uint8Array data. Should succeed.',
     ),
