+++ ldk/javascript/src/document/types.ts
@@ -1,0 +1,21 @@
+
+export interface Workbook {
+    worksheets: Worksheet[];
+  }
+  
+  export interface Worksheet {
+    hidden: boolean;
+    hiddenColumns: number[];
+    hiddenRows: number[];
+    name: string;
+    rows: Row[];
+  }
+  
+  export interface Row {
+    cells: Cell[];
+  }
+  
+  export interface Cell {
+    value: string;
+  }
+
