+++ ldk/javascript/src/types/oliveHelps/document.d.ts
